Added error handling and added few more
